consumer/requirements.txt
aiokafka>=0.11
python-dotenv>=1.0
anthropic>=0.45
pydantic>=2.0
redis>=5.0
